+++ warp/config.py
@@ -16,7 +16,8 @@ verify_cuda = False     # if true will check CUDA errors after each kernel launc
 print_launches = False  # if true will print out launch information
 
 mode = "release"
-verbose = False
+verbose = False         # print extra informative messages
+quiet = False           # suppress all output except errors and warnings
 
 host_compiler = None    # user can specify host compiler here, otherwise will attempt to find one automatically